ability to set different click for nicklist
