+++ book/src/configuration/buffer.md
@@ -138,6 +138,38 @@ Overwrite nicklist width in pixels.
 width = 150
 ```
 
+#### `click`
+
+Click action for when interaction with nicknames.
+
+##### `action`
+
+The action to perform.  
+- `"open-query"`: Open a query with the User
+- `"insert-nickname"`: Inserts the nickname into text input
+
+```toml
+# Type: string
+# Values: "open-query", "insert-nickname"
+# Default: "open-query"
+
+[buffer.channel.nicklist.click]
+action = "open-query"
+```
+
+##### `interaction`
+
+The interaction to perform a action.  
+
+```toml
+# Type: string
+# Values: "single-click", "double-click"
+# Default: "double-click"
+
+[buffer.channel.nicklist.click]
+interaction = "double-click"
+```
+
 ### `[buffer.channel.topic]`
 
 Topic settings within a channel buffer.
