+++ data/src/config/buffer.rs
@@ -35,6 +35,31 @@ pub struct Buffer {
     pub emojis: Emojis,
 }
 
+#[derive(Debug, Clone, Deserialize, Default)]
+#[serde(rename_all = "kebab-case")]
+pub struct NicknameClick {
+    #[serde(default)]
+    pub action: NicknameClickAction,
+    #[serde(default)]
+    pub interaction: NicknameClickInteraction,
+}
+
+#[derive(Debug, Clone, Deserialize, Default)]
+#[serde(rename_all = "kebab-case")]
+pub enum NicknameClickAction {
+    #[default]
+    OpenQuery,
+    InsertNickname,
+}
+
+#[derive(Debug, Clone, Deserialize, Default)]
+#[serde(rename_all = "kebab-case")]
+pub enum NicknameClickInteraction {
+    SingleClick,
+    #[default]
+    DoubleClick,
+}
+
 #[derive(Debug, Clone, Deserialize)]
 pub struct Emojis {
     #[serde(default = "default_bool_true")]
